Avoid running tests in parallel
This prevents a number of simulator clones from being run.

Edit the Test scheme, select the small arrow beside the Birthdays unit
test targe which opens the auto-generated Test Plan. Click the
Options... button and unselect the option:
  Execute in parallel (if possible)

This also creates a Test Plan when we modify the test scheme.